+++ test/Braintree.Tests.Integration/PaymentMethodIntegrationTest.cs
@@ -84,6 +84,8 @@ namespace Braintree.Tests.Integration
             Assert.IsNotNull(paymentMethodResult.Target.ImageUrl);
             Assert.AreEqual(result.Target.Id, paymentMethodResult.Target.CustomerId);
             Assert.IsInstanceOf(typeof(PayPalAccount), paymentMethodResult.Target);
+            PayPalAccount paypalAccount = (PayPalAccount) paymentMethodResult.Target;
+            Assert.IsNotNull(paypalAccount.PayerId);
         }
 
         [Test]
@@ -125,6 +127,8 @@ namespace Braintree.Tests.Integration
             Assert.IsNotNull(paymentMethodResult.Target.ImageUrl);
             Assert.AreEqual(result.Target.Id, paymentMethodResult.Target.CustomerId);
             Assert.IsInstanceOf(typeof(PayPalAccount), paymentMethodResult.Target);
+            PayPalAccount paypalAccount = (PayPalAccount) paymentMethodResult.Target;
+            Assert.IsNotNull(paypalAccount.PayerId);
         }
 
         [Test]